chore: type correction

+++ src/views/environment.py
@@ -43,7 +43,7 @@ class EnvironmentSimulation(ApiBaseView):
     initial_east: Optional[float] = None
     initial_hemisphere: Optional[str] = None
     initial_ew: Optional[str] = None
-    max_expected_height: Optional[int] = None
+    max_expected_height: Optional[float] = None
     date: Optional[datetime] = Field(default_factory=_default_future_datetime)
     local_date: Optional[datetime] = Field(
         default_factory=_default_future_datetime
